Adopt canonical types throughout (i16/void), align grammar and tests; rename leftover SomeCompiler profile to RetroSharp; re-enable semantic sample suite with updated snapshots (#25)

+++ src/RetroSharp.SemanticAnalysis/IntSymbol.cs
@@ -1,6 +1,6 @@
 ﻿namespace RetroSharp.SemanticAnalysis;
 
-public record IntSymbol() : SymbolType("int")
+public record IntSymbol() : SymbolType("i16")
 {
     public static readonly IntSymbol Instance = new();
     public override string ToString() => base.ToString();
