+++ src/RetroSharp.SemanticAnalysis/IntType.cs
@@ -1,6 +1,6 @@
 ﻿namespace RetroSharp.SemanticAnalysis;
 
-public record IntType() : SymbolType("int")
+public record IntType() : SymbolType("i16")
 {
     public static readonly IntType Instance = new();
     public override string ToString() => base.ToString();
